perf(format): shrink ParseEvent from 96 to 48 bytes (#1949)
## Summary

Optimize core event types for better memory efficiency:

| Type | Before | After | Reduction |
|------|--------|-------|-----------|
| Span | 16 bytes | 8 bytes | 50% |
| FieldKey | 72 bytes | 24 bytes | 67% |
| ParseEventKind | 80 bytes | 32 bytes | 60% |
| ParseEvent | 96 bytes | 48 bytes | 50% |

### Changes

**Span** now uses `u32` for offset and length instead of `usize`. This
supports files up to 4GB which is sufficient for all practical use
cases.

**FieldKey** is now an enum with two variants:
- `Name(Cow<str>)`: common case for JSON/YAML/TOML (24 bytes)
- `Full(Box<FullFieldKey>)`: for Styx with doc/tag metadata (8 bytes
pointer)

### Benchmark Results

| Benchmark      | Before (PR #1948) | After    | Change  |
|----------------|-------------------|----------|---------|
| facet_json     | ~14.4 ms          | ~13.7 ms | **-5%** |
| facet_json_str | ~14.3 ms          | ~13.6 ms | **-5%** |
| serde_json     | 1.7 ms            | 1.7 ms   | -       |

## Test plan

- [x] All workspace tests pass (2027 tests)
- [x] Benchmarks show ~5% improvement

+++ facet-json/src/parser.rs
@@ -159,7 +159,7 @@ impl<'de, const TRUSTED_UTF8: bool> JsonParser<'de, TRUSTED_UTF8> {
                 .map_err(scan_error_to_parse_error)?;
         }
 
-        self.state.last_token_start = spanned.span.offset;
+        self.state.last_token_start = spanned.span.offset as usize;
         self.state.scanner_pos = self.scanner.pos();
 
         let kind = match spanned.token {
@@ -369,7 +369,7 @@ impl<'de, const TRUSTED_UTF8: bool> JsonParser<'de, TRUSTED_UTF8> {
             .scanner
             .next_token(self.input)
             .map_err(scan_error_to_parse_error)?;
-        let start = first.span.offset;
+        let start = first.span.offset as usize;
         self.state.scanner_pos = self.scanner.pos();
 
         match first.token {
@@ -783,7 +783,7 @@ impl<'de, const TRUSTED_UTF8: bool> FormatParser<'de> for JsonParser<'de, TRUSTE
                 .scanner
                 .next_token(self.input)
                 .map_err(scan_error_to_parse_error)?;
-            let start = first.span.offset;
+            let start = first.span.offset as usize;
             self.state.scanner_pos = self.scanner.pos();
 
             // Skip the rest of the value if it's a container
